+++ alacritty_terminal/src/config/colors.rs
@@ -79,6 +79,10 @@ pub struct CursorColors {
     pub text: Option<Rgb>,
     #[serde(deserialize_with = "failure_default")]
     pub cursor: Option<Rgb>,
+    #[serde(deserialize_with = "failure_default")]
+    pub keyboard_motion_text: Option<Rgb>,
+    #[serde(deserialize_with = "failure_default")]
+    pub keyboard_motion_cursor: Option<Rgb>,
 }
 
 #[serde(default)]